Refactor yolov7 (#1153)
* use cmake 3.10

* refactor yolov7 code

* refactor yolov7 parse_args

* remove anchor definition
