+++ yolov7/include/logging.h
